Refactor. change extract fns to utils
